requirements.txt
pandas
numpy
scipy
plotly
yfinance
sqlalchemy
fastapi
uvicorn[standard]
pyyaml
pyportfolioopt